--- src/components/Editor/EditForm.tsx
@@ -1,7 +1,0 @@
-import styled from "@emotion/styled";
-import { css } from "@emotion/react";
-import { PALLETS_LIGHT } from "../../constants";
-
-export const BottomMenu = () => {
-  return <div></div>;
-};
